backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.4.0
pydantic==1.10.15
qrcode==7.4.2
Pillow==10.4.0
reportlab==4.2.5
python-multipart==0.0.12
